requirements.txt
# Core dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.21.0
pandas>=1.5.0

# Machine Learning
tensorflow>=2.12.0
scikit-learn>=1.2.0

# UI and Visualization
streamlit>=1.25.0
matplotlib>=3.6.0
seaborn>=0.11.0
plotly>=5.15.0

# Data Processing
Pillow>=9.0.0
scipy>=1.9.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.64.0
colorama>=0.4.0

# Development and Testing
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: For better performance
# torch==2.0.1
# torchvision==0.15.2